Lots of changes.
Added burn dust and burn projectiles, experimenting with a new style of aura.
Added Cerberus, a playground shield accessory.
Added sprite for FireEssence.
Added Elemental Relics, added Chaos Catalyst, added Lunar Sigils
Removed quotes from item descriptions.
Consolidated essences to their own files for organization, and to allow easier glowmask manipulation and ease of individual code changes.
Added Starfiber.
Attempting to make our AfterImage method more reliable.
Added more parameters for our aura generation method.
Added classes to utils for aura type and aura style to be used in conjunction with our aura generation method.
Updated the efficiency of our elemental damage number helper.

+++ UI/BossHealth.cs
@@ -72,10 +72,6 @@ namespace ProvidenceMod.UI
     public override void OnInitialize()
     {
       // Here we initialize the area and set our variables
-      // Never use odd numbers for placement, or sprite dimensions
-      // For some odd reason, Terraria does not like odd numbers when it comes to UI 
-      // The first thing you'll notice when you use them is that your UI looks nowhere near as crisp as vanilla UI
-      // My assumption is that this is because of how it renders the UI, it may blur the edges when the dimensions are uneven
       area = new UIElement();
       // The first valoe is the pixel count fron the left or top, the second value is the percentage in case that's easier to use
       area.Left.Set(Main.screenWidth - 549, 0f);
@@ -196,8 +192,7 @@ namespace ProvidenceMod.UI
         {
           bossHealth = bossNPC.life;
         }
-        // Don't change this please, it works, Roslynator is wack
-        quotient = ((float)bossHealth) / ((float)bossHealthMax);
+        quotient = ((float)bossHealth) / bossHealthMax;
         if (!arraySet)
         {
           lifeArray[2] = lifeArray[1];
